requirements.txt
google-generativeai
python-dotenv
streamlit
PyPDF2